requirements.txt
# Core lab pipeline
openai>=1.40.0
python-dotenv>=1.0.0

# Retrieval / NLP (M1–M3)
sentence-transformers>=3.0.0
torch>=2.1.0
qdrant-client>=1.7.0
rank-bm25>=0.2.2
underthesea>=6.8.0

# Evaluation (M4)
ragas>=0.4.0
datasets>=2.14.0
langchain-openai>=0.1.0

# Guardrails (optional Presidio — có fallback regex)
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
# Presidio Analyzer (English): python -m spacy download en_core_web_sm

# Dev
pytest>=8.0.0
ruff>=0.4.0
